+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2 (Mar 21, 2020)
+
+* Change the card set to neutral so it works with Rook(Probably).
+* Change "Blackmail" to common from uncommon.
+* Added "Fake Promise", a new card which allows you to negotiate better if you have the shills.
+
 ## 0.0.1 (Mar 21, 2020)
 
 * Initial commit with 4 new cards.